Source the ROS Gazebo system setup file
The ROS workspace setup file does not source the ROS Gazebo system setup
file.

This adds the sourcing for the ROS Gazebo system setup file. It drops
the workaround that sets the gazebo-11 system's directory. The sourcing
for the ROS Gazebo system setup file was the missing bits.

Note: This exports the following additional Gazebo variable to the
environment:
 - GAZEBO_MASTER_URI=http://localhost:11345
 - GAZEBO_PLUGIN_PATH=/usr/share/gazebo/../../lib/x86_64-linux-gnu/gazebo-11/plugins:
 - GAZEBO_MODEL_DATABASE_URI=http://models.gazebosim.org
 - GAZEBO_RESOURCE_PATH=/usr/share/gazebo/../../share/gazebo-11:
 - GAZEBO_MODEL_PATH=/usr/share/gazebo/../../share/gazebo-11/models:
 - OGRE_RESOURCE_PATH=/usr/lib/x86_64-linux-gnu/OGRE-1.9.0

Also, the LD_LIBRARY_PATH lists the extra Gazebo plugin directory
/opt/ros/humble/lib:/usr/lib/x86_64-linux-gnu/gazebo-11/plugins:
 -LD_LIBRARY_PATH=/usr/lib/x86_64-linux-gnu/gazebo-11/plugins:/opt/ros/humble/opt/rviz_ogre_vendor/lib:/opt/ros/humble/lib/x86_64-linux-gnu:/opt/ros/humble/lib
 +LD_LIBRARY_PATH=/usr/lib/x86_64-linux-gnu/gazebo-11/plugins:/opt/ros/humble/opt/rviz_ogre_vendor/lib:/opt/ros/humble/lib/x86_64-linux-gnu:/opt/ros/humble/lib:/usr/share/gazebo/../../lib/x86_64-linux-gnu/gazebo-11/plugins:

[1]: https://classic.gazebosim.org/tutorials?tut=ros2_installing#Introduction