Source the ROS system setup file
The ROS workspace setup file does not source the ROS system setup file.

This adds the sourcing for the ROS system setup file.

Note: This seems to be unnecessary as the environment is the same if
sourcing the workspace setup files only, and if sourcing both setup
files (i.e. system and workspace).

[1]: https://docs.ros.org/en/humble/Tutorials/Beginner-CLI-Tools/Configuring-ROS2-Environment.html#source-the-setup-files