robot_control: sets GAZEBO_MODEL_PATH
This appends the robot_control's models directory to the environ
GAZEBO_MODEL_PATH, using the action AppendEnvironmentVariable.

Note: It sets the gazebo-11 system's directories too, to workaround some
missing bits.

Note: This is unnecessary as the robot_controls's resource directory is
appended to the environment GAZEBO_RESOURCE_PATH already. That resource
directory contains the models directory and extra resources that are
necessary to ROS Gazebo.