robot_description: gazebo_move: fix path to URDF Xacro file
The package robot_moveit_config does not contains the URDF Xacro file.

The URDF Xacro file r5a_v_ros.urdf.xacro is located in the directory
urdf/ of the robot_description package.

This fixes the path to the URDF Xacro file to address the error below,
it reuses the variable xacro_file containing the full path to the file.

Fixes:

	WARNING:root:"File /home/gportay/src/abmi-software/Robot5A-Simulation/install/robot_moveit_config/share/robot_moveit_config/config/r5a_v_ros.urdf.xacro doesn't ex
	WARNING:root:The robot description will be loaded from /robot_description topic